Refactoring SFTP to use the new minio/pkg inplementation
